Azure, Wrap, Test: Use python version 3.9 (latest) for Mac
There was a failure, where in Mac, the desired python 3.8 was not getting used
in CMake (but 3.9).

And test in Release mode in Windows.
I cannot avoid CMake appending `d` to the DGtal libraries (dgtald.lib),
tried setting CMAKE_DEBUG_POSTFIX to the empty string `""`, but not
working.

I think it's all right to test with different settings in different
platforms.